server/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
supabase==2.18.1
httpx
python-multipart
